update to use app router
