+++ package.json
@@ -4,7 +4,7 @@
   "private": true,
   "scripts": {
     "dev": "next dev",
-    "build": "next build && next export",
+    "build": "next build",
     "start": "next start",
     "enableWebFrameworks": "firebase experiments:enable webframeworks",
     "lint": "next lint --dir src",
@@ -12,9 +12,8 @@
     "format": "prettier --write --ignore-path .gitignore './**/*.{js,jsx,ts,tsx,json}'"
   },
   "dependencies": {
-    "@next/font": "13.1.4",
     "@types/node": "18.11.18",
-    "@types/react": "18.0.27",
+    "@types/react": "^18.3.1",
     "@types/react-dom": "18.0.10",
     "autoprefixer": "^10.4.13",
     "eslint": "8.32.0",
@@ -22,7 +21,6 @@
     "firebase-admin": "^11.10.1",
     "moment": "^2.29.4",
     "next": "^14.2.3",
-    "next-seo": "^6.1.0",
     "postcss": "^8.4.21",
     "postcss-nested": "^6.0.0",
     "react": "^18.3.1",
@@ -37,5 +35,8 @@
     "eslint-plugin-tailwindcss": "^3.8.2",
     "prettier": "^2.8.4",
     "prettier-plugin-tailwindcss": "^0.2.2"
+  },
+  "resolutions": {
+    "@types/react": "^18.3.1"
   }
 }
