+++ src/components/GoogleAnalytics.tsx
@@ -1,0 +1,41 @@
+'use client'
+
+import { usePathname, useSearchParams } from 'next/navigation'
+import Script from 'next/script'
+import { useEffect } from 'react'
+
+import { existsGaId, GA_MEASUREMENT_ID, pageview } from '@/lib/gtag'
+
+const GoogleAnalytics = () => {
+  const pathname = usePathname()
+  const searchParams = useSearchParams()
+
+  useEffect(() => {
+    if (!existsGaId) {
+      return
+    }
+    const path = pathname + searchParams.toString()
+    pageview(path)
+  }, [pathname, searchParams])
+
+  return (
+    <>
+      <Script
+        src={`https://www.googletagmanager.com/gtag/js?id=${GA_MEASUREMENT_ID}`}
+        strategy='lazyOnload'
+      />
+      <Script id='gtag-init' defer strategy='afterInteractive'>
+        {`
+          window.dataLayer = window.dataLayer || [];
+          function gtag(){dataLayer.push(arguments);}
+          gtag('js', new Date());
+          gtag('config', '${GA_MEASUREMENT_ID}', {
+            page_path: window.location.pathname,
+          });
+        `}
+      </Script>
+    </>
+  )
+}
+
+export default GoogleAnalytics
