+++ src/components/HeroLetter.tsx
@@ -5,8 +5,13 @@ const HeroLetter = () => {
   return (
     <div className=''>
       <div className='relative aspect-video'>
-        <Image src='/img/hero-letter-bg.jpg' fill alt='' className='aspect-video object-cover' />
-        {/* </div> */}
+        <Image
+          src='/img/hero-letter-bg.jpg'
+          fill
+          alt=''
+          priority={false}
+          className='aspect-video object-cover'
+        />
         <Image
           src='/img/hero-letter-cover.svg'
           width={1280}
@@ -18,7 +23,7 @@ const HeroLetter = () => {
           src='/img/hero-fb-mark.svg'
           width={192}
           height={192}
-          className='absolute right-[5%] bottom-[10%] z-20 w-[15%] rotate-[-10deg]'
+          className='absolute bottom-[10%] right-[5%] z-20 w-[15%] rotate-[-10deg]'
           alt=''
         />
       </div>
