--- src/hooks/usePageView.ts
@@ -1,26 +1,0 @@
-import { useRouter } from 'next/router'
-import { useEffect } from 'react'
-
-import { existsGaId, pageview } from '@/lib/gtag'
-
-const usePageView = () => {
-  const router = useRouter()
-
-  useEffect(() => {
-    if (!existsGaId) {
-      return
-    }
-
-    const handleRouteChange = (path: string) => {
-      pageview(path)
-    }
-
-    router.events.on('routeChangeComplete', handleRouteChange)
-
-    return () => {
-      router.events.off('routeChangeComplete', handleRouteChange)
-    }
-  }, [router.events])
-}
-
-export default usePageView
